chore/refatora_producers
